requirements.txt
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
cryptography==3.1
dnspython==2.0.0
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
pyOpenSSL==19.1.0
six==1.15.0
Werkzeug==1.0.1
